Move code in run.py into cli.py
